+++ lib/src/web_storage_manager.dart
@@ -6,6 +6,8 @@ import 'package:flutter/services.dart';
 import 'types.dart';
 
 ///Class that implements a singleton object (shared instance) which manages the web storage used by WebView instances.
+///On Android, it is implemented using [WebStorage](https://developer.android.com/reference/android/webkit/WebStorage.html).
+///On iOS, it is implemented using [WKWebsiteDataStore.default()](https://developer.apple.com/documentation/webkit/wkwebsitedatastore).
 ///
 ///**NOTE for iOS**: available from iOS 9.0+.
 class WebStorageManager {
